move traceroute to its own file, and a gauge metric
