+++ internal/scraper/config.go
@@ -1,0 +1,10 @@
+package scraper
+
+import (
+	bbeconfig "github.com/prometheus/blackbox_exporter/config"
+)
+
+type ConfigModule struct {
+	bbeconfig.Module
+	Traceroute TracerouteProbe `yaml:"traceroute,omitempty"`
+}
